Centralise line chart story config and move new glyph story
