+++ projects/js-packages/charts/src/components/line-chart/stories/index.stories.tsx
@@ -1,7 +1,4 @@
-import { GlyphDiamond, GlyphStar } from '@visx/glyph';
 import React from 'react';
-import { ThemeProvider, jetpackTheme, wooTheme } from '../../../providers/theme';
-import { DefaultGlyph } from '../../shared/default-glyph';
 import LineChart from '../line-chart';
 import { lineChartStoryArgs, lineChartMetaArgs } from './config';
 import largeValuesData from './large-values-sample';
@@ -9,94 +6,9 @@ import sampleData from './sample-data';
 import webTrafficData from './site-traffic-sample';
 import type { Meta, StoryFn, StoryObj } from '@storybook/react';
 
-const customStorybookTheme = {
-	...jetpackTheme,
-	glyphs: [
-		props => React.createElement( DefaultGlyph, { ...props, key: props.key } ),
-		props =>
-			React.createElement( GlyphStar, {
-				key: props.key,
-				top: props.y,
-				left: props.x,
-				size: props.size * props.size,
-				fill: props.color,
-			} ),
-		props =>
-			React.createElement( GlyphDiamond, {
-				key: props.key,
-				top: props.y,
-				left: props.x,
-				size: props.size * props.size,
-				fill: props.color,
-			} ),
-	],
-};
-
-const THEME_MAP = {
-	default: undefined,
-	jetpack: jetpackTheme,
-	woo: wooTheme,
-	customStorybook: customStorybookTheme,
-};
-
 const meta: Meta< typeof LineChart > = {
 	...lineChartMetaArgs,
 	title: 'JS Packages/Charts/Types/Line Chart',
-	component: LineChart,
-	parameters: {
-		layout: 'centered',
-	},
-	decorators: [
-		( Story, { args } ) => {
-			const theme = THEME_MAP[ args.themeName ];
-
-			return (
-				<ThemeProvider theme={ theme }>
-					<div
-						style={ {
-							resize: 'both',
-							overflow: 'auto',
-							padding: '2rem',
-							width: '800px',
-							maxWidth: '1200px',
-							border: '1px dashed #ccc',
-							display: 'inline-block',
-						} }
-					>
-						<Story />
-					</div>
-				</ThemeProvider>
-			);
-		},
-	],
-	argTypes: {
-		themeName: {
-			control: 'select',
-			options: [ 'default', 'jetpack', 'woo', 'customStorybook' ],
-			defaultValue: 'default',
-		},
-		maxWidth: {
-			control: {
-				type: 'number',
-				min: 100,
-				max: 1200,
-			},
-		},
-		aspectRatio: {
-			control: {
-				type: 'number',
-				min: 0,
-				max: 1,
-			},
-		},
-		resizeDebounceTime: {
-			control: {
-				type: 'number',
-				min: 0,
-				max: 10000,
-			},
-		},
-	},
 } satisfies Meta< typeof LineChart >;
 
 export default meta;
@@ -356,78 +268,6 @@ BrokenLine.parameters = {
 		description: {
 			story: 'Demonstrates the option of setting a seriesLineStyle to a dash array.',
 		},
-	},
-};
-
-export const WithStartGlyphs: StoryObj< typeof LineChart > = Template.bind( {} );
-WithStartGlyphs.args = {
-	...Default.args,
-	withStartGlyphs: true,
-};
-
-export const WithCustomGlyph: StoryObj< typeof LineChart > = Template.bind( {} );
-WithCustomGlyph.args = {
-	...Default.args,
-	showLegend: true,
-	withStartGlyphs: true,
-	withLegendGlyph: true,
-	renderGlyph: ( { color, size, x, y } ) => {
-		return <GlyphStar top={ y } left={ x } size={ size * size } fill={ color } />;
-	},
-	glyphStyle: {
-		radius: 10,
-	},
-};
-
-const CustomStarGlyph = ( { color, size, x, y } ) => {
-	const hasXY = typeof x === 'number' && typeof y === 'number' && ( x !== 0 || y !== 0 );
-	const groupProps = hasXY ? { transform: `translate(${ x }, ${ y })` } : {};
-	return (
-		<g { ...groupProps }>
-			<svg
-				xmlns="http://www.w3.org/2000/svg"
-				width={ size * 2 }
-				height={ size * 2 }
-				viewBox="0 0 24 24"
-				style={ { overflow: 'visible', pointerEvents: 'none' } }
-			>
-				<path
-					d="M20.84 4.61a5.5 5.5 0 0 0-7.78 0L12 5.67l-1.06-1.06a5.5 5.5 0 0 0-7.78 7.78l1.06 1.06L12 21.23l7.78-7.78 1.06-1.06a5.5 5.5 0 0 0 0-7.78z"
-					fill={ color }
-					stroke={ color }
-					strokeWidth="2"
-					strokeLinecap="round"
-					strokeLinejoin="round"
-					transform="translate(-12, -12)"
-				/>
-			</svg>
-		</g>
-	);
-};
-
-export const WithCustomSvgGlyph: StoryObj< typeof LineChart > = Template.bind( {} );
-WithCustomSvgGlyph.args = {
-	...Default.args,
-	showLegend: true,
-	withStartGlyphs: true,
-	withLegendGlyph: true,
-	renderGlyph: ( { color, size, x, y } ) => (
-		<CustomStarGlyph color={ color } size={ size } x={ x } y={ y } />
-	),
-	glyphStyle: {
-		radius: 8,
-	},
-};
-
-export const WithCustomGlyphsPerDataPoint: StoryObj< typeof LineChart > = Template.bind( {} );
-WithCustomGlyphsPerDataPoint.args = {
-	...Default.args,
-	showLegend: true,
-	withStartGlyphs: true,
-	withLegendGlyph: true,
-	themeName: 'customStorybook', // Mock prop used to switch the rendered theme in the storybook.
-	glyphStyle: {
-		radius: 8,
 	},
 };
